requirements.txt
# Jen AI Assistant - Python Dependencies
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database connectivity
pyodbc==4.0.39

# AI and HTTP requests
requests==2.31.0
openai==1.3.7

# Environment and configuration
python-dotenv==1.0.0

# WebSocket support
websockets==12.0

# Logging and utilities
python-multipart==0.0.6

# Audio processing (for future enhancements)
pydub==0.25.1

# Date/time utilities
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# CORS and middleware
starlette==0.27.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Twilio integration
twilio==8.10.0